requirements.txt
stomp.py
jinja2
ansible<2.0
